Ship source-aligned landing CMS and image previews

+++ packages/panel/src/Support/Contrast.php
@@ -86,9 +86,9 @@ final class Contrast
         }
 
         return [
-            hexdec(substr($hex, 0, 2)),
-            hexdec(substr($hex, 2, 2)),
-            hexdec(substr($hex, 4, 2)),
+            (int) hexdec(substr($hex, 0, 2)),
+            (int) hexdec(substr($hex, 2, 2)),
+            (int) hexdec(substr($hex, 4, 2)),
         ];
     }
 }
